Drop Virtualization SIG IRC meetings
The CentOS Virtualization Special Interest Group
will not be holding regular meetings for the time being.
Instead, all discussions and collaborations should take place on our
[mailing list](https://lists.centos.org/hyperkitty/list/virt@lists.centos.org/).

If a meeting is needed for a specific topic, members can request one through the mailing list,
and we will organize it as necessary.

We appreciate your participation and look forward to engaging discussions online.

Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>